+++ ImageFeed/ImagesListService/ImagesListService.swift
@@ -1,48 +1,5 @@
 import Foundation
 
-struct LikeResult: Codable {
-    let photo: PhotoResult
-}
-
-struct PhotoResult: Codable {
-    let id: String
-    let width: Int
-    let height: Int
-    let createdAt: String?
-    let description: String?
-    let likedByUser: Bool
-    let urls: UrlsResult
-    
-    enum CodingKeys: String, CodingKey {
-        case id
-        case width
-        case height
-        case description
-        case urls
-        case createdAt = "created_at"
-        case likedByUser = "liked_by_user"
-    }
-}
-
-struct UrlsResult: Codable {
-    let thumb: String
-    let full: String
-}
-
-struct Photo: Equatable {
-    let id: String
-    let size: CGSize
-    let createdAt: Date?
-    let welcomeDescription: String?
-    let thumbImageURL: String
-    let largeImageURL: String
-    let isLiked: Bool
-    
-    static func == (lhs: Photo, rhs: Photo) -> Bool {
-        return lhs.id == rhs.id
-    }
-}
-
 final class ImagesListService {
     static let shared = ImagesListService()
     static let didChangeNotification = Notification.Name(rawValue: "ImagesListServiceDidChange")
@@ -55,6 +12,8 @@ final class ImagesListService {
     private var lastLoadedPage: Int?
 
     func fetchPhotosNextPage() {
+        guard photosTask == nil else { return }
+        
         var nextPage: Int
         if let lastLoadedPage = lastLoadedPage {
             nextPage = lastLoadedPage + 1
@@ -148,8 +107,8 @@ final class ImagesListService {
                     isLiked: !photo.isLiked
                 )
                 self.photos = self.photos.withReplaced(itemAt: index, newValue: newPhoto)}
-        case .failure:
-            break
+        case .failure(let error):
+            assertionFailure("\(error)")
         }
     }
 }
