+++ ImageFeed/ImagesListService/Models.swift
@@ -1,0 +1,44 @@
+import Foundation
+
+struct LikeResult: Codable {
+    let photo: PhotoResult
+}
+
+struct PhotoResult: Codable {
+    let id: String
+    let width: Int
+    let height: Int
+    let createdAt: String?
+    let description: String?
+    let likedByUser: Bool
+    let urls: UrlsResult
+    
+    enum CodingKeys: String, CodingKey {
+        case id
+        case width
+        case height
+        case description
+        case urls
+        case createdAt = "created_at"
+        case likedByUser = "liked_by_user"
+    }
+}
+
+struct UrlsResult: Codable {
+    let thumb: String
+    let full: String
+}
+
+struct Photo: Equatable {
+    let id: String
+    let size: CGSize
+    let createdAt: Date?
+    let welcomeDescription: String?
+    let thumbImageURL: String
+    let largeImageURL: String
+    let isLiked: Bool
+    
+    static func == (lhs: Photo, rhs: Photo) -> Bool {
+        return lhs.id == rhs.id
+    }
+}
